Massively-improve "backtick names" thanks to @xeno-by.
This commit introduces a StringParser class, which can run the
Scala parser on small snippets. Using this, we can avoid a lot
of the pain of manual parsing/tree building, and be sure that
our synthetic trees are the same as we'd get from the parser.

Note that this commit only works for 2.11. We may need to use
scala_2.10 and scala_2.11 source directories to support both
versions, and it's likely that we'll need to drop 2.9 support
to merge this branch (no big loss in my opinion).