Add code to do error-checking within Lambda[...].
Previously only the shape of the type argument to Lambda was
being used. This meant that instead of using a Function1, you
could use a Tuple2 and get the same results:

    Lambda[a => Either[Int, a]] // is ok
    Lambda[(a, Either[Int, a])] // is ok, should be error

    type L[A] = (A => A)
    Lambda[L[String]] // does not do what author intended

This commit checks the application occuring within Lambda[...]
to be sure that the target is a scala.FunctionN, where N >= 1.

This fixes #18.