Improve the efficiency of kind-projector.
This fixes an issue raised by tpolecat where kind-projector
caused scalac to seem to hang when used with large hlist
types and shapeless. Kind-projector syntax was not being
used, so it seemed like the tree transformers were somehow
taking too long.

The problem ended up being that there was a quadratic
slowdown when kp parsed recursive types like hlists.
The solution was to cache (memoize) the `Tree => Tree`
transformation, to ensure that we don't recompute
these trees many times.

It would be nice if we could somehow clear this cache
between source files, or top-level classes, to avoid
using huge amounts of memory during compilation. I'm
going to explore that now.