Add a section on polymorphic lambdas to the README.
This PR is an interesting exercise -- it has forced me to test my understanding
of the feature by trying to write documentation for it. It's possible that we
need another section for polymorphic lambda caveats or errata, but I'd prefer
to keep the initial section relatively high-level.

I plan to release this in 0.9.0 as soon as we feel like we have adequate
documentation to describe what is going on.

Review by @paulp.