Syntax for polymorphic values.
    ν[T].m[A, B[_], ...](e)

is rewritten to

    new T { def m[A, B[_], ...] = e }

(`ν` is the lowercase Greek letter "Nu" (/njuː/) and stands for `new`.)

In case `m` has just one *-kinded type parameter that is
is not referenced from `e`, it can be omitted:

    ν[T].m(e)

is rewritten to

    new T { def m[A] = e }

where `A` is a fresh name.

If the method name is `apply`, it can be omitted:

    ν[T][A, B[_], ...](e)

is rewritten to

    new T { def apply[A, B[_], ...] = e }

The previous two shortcuts can be combined:

    ν[T](e)

is rewritten to

    new T { def apply[A] = e }

where `A` is a fresh name.