Remove deprecated `?` syntax
Scala 2.13.6 and 2.12.14 will interpret `?` as a wildcard when using the
`-Xsource:3` flag (cf https://github.com/scala/scala/pull/9560)). This
means that the old kind-projector syntax will no longer work, so it
seems like a good time to remove it. This will also allow us to compile
more of the community-build with `-Xsource:3` enabled (cf
https://github.com/scala/scala-dev/issues/769).

Sincet this is a breaking change, we also bump the version to
0.12.0-SNAPSHOT.